LPS-63806 Move the helper method to the parent class so we can invoke it for the required places
